+++ includes/class-solrpower-sync.php
@@ -734,16 +734,27 @@ class SolrPower_Sync {
 	 * when the plugin posts data to Solr, but your Solr instance will need
 	 * to have a cron job enabled that does a hard commit on a regular basis.
 	 *
-	 * To disable commiting to Solr, add the following to your wp-config.php
+	 *
+	 * To enable commiting to Solr, add the following to your wp-config.php
 	 *
 	 * <code>
-	 * define( 'SOLRPOWER_DISABLE_AUTOCOMMIT', true );
+	 * define( 'SOLRPOWER_DISABLE_AUTOCOMMIT', false );
 	 * </code>
+	 *
+	 * As of 2.5.0, the default behavior is to not auto-commit.
+	 * Yes, this function could be more concise, but the double-negative logic
+	 * can be confusing so making the logic as explicit as possible.
 	 *
 	 * @see https://cwiki.apache.org/confluence/display/solr/UpdateXmlMessages#UpdateXmlMessages-%22commit%22and%22optimize%22
 	 * @return bool Whether to commit immediately when writing site data to Solr.
 	 */
 	function should_commit() {
-		return ! ( defined( 'SOLRPOWER_DISABLE_AUTOCOMMIT' ) && SOLRPOWER_DISABLE_AUTOCOMMIT );
+		// Auto-commit is explicitly enabled by not disabling it 🤮.
+		if ( defined( 'SOLRPOWER_DISABLE_AUTOCOMMIT' ) && ! SOLRPOWER_DISABLE_AUTOCOMMIT ) {
+			return true;
+		}
+
+		// Do not autocommit, as cron will take care of it.
+		return false;
 	}
 }